playstore-production-automation/requirements.txt
anthropic>=0.34.0
requests>=2.31.0
pyyaml>=6.0.1
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0